+++ backend/src/models/article.rs
@@ -21,7 +21,7 @@ pub struct Metadata {
 pub struct Article {
     pub slug: String,
     pub metadata: Metadata,
-    pub version: u32,
+    pub version: u64,
     pub updated_at: DateTime<Utc>,
     #[serde(skip_serializing)]
     pub file_path: String,
